+++ app/Http/Controllers/Admin/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\UserRequest;
 use App\Models\User;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Inertia\Inertia;
@@ -12,19 +13,24 @@ use Illuminate\Validation\Rules;
 
 class UserController extends Controller
 {
+    use AuthorizesRequests;
+
     public function index()
     {
+        $this->authorize('viewAny', User::class);
         $users = User::role('user')->get();
         return Inertia::render('Admin/Users/Index', ['users' => $users]);
     }
 
     public function create()
     {
+        $this->authorize('create', User::class);
         return Inertia::render('Admin/Users/Form');
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', User::class);
         $request->validate([
             'name' => 'required|string|max:255',
             'email' => 'required|string|lowercase|email|max:255|unique:'.User::class,
@@ -39,11 +45,15 @@ class UserController extends Controller
 
         $user->assignRole('user');
 
+        $permissions = ['can-list-posts', 'can-create-posts', 'can-edit-posts', 'can-delete-posts', 'can-restore-posts'];
+        $user->givePermissionTo($permissions);
+
         return redirect()->route('admin.users.index');
     }
 
     public function destroy(User $user)
     {
+        $this->authorize('delete', $user);
         $user->delete();
         return redirect()->route('admin.users.index');
     }
